refactor: share shortcut settings and isolate double-tap controls
